Add DOM Content Loaded event listener to index.js
It's to make sure the content is loaded before checking and editing any of it. Usually it should be fine since the script tag is at the bottom of the file anyway, but you can't be too sure. This doesn't have to be for the whole script, just what needs the HTML to be loaded before doing anything. For this, the first line already needs to find DOM elements.